Refactor: Overhaul project structure, migrate to TypeScript, and improve UI
- Migrated all .jsx files to .tsx
- Restructured folders for better organization
- Refactored Dashboard and AdminPanel into smaller components
- Created reusable StatCard and Layout components
- Applied a consistent blue color scheme
- Fixed all TypeScript errors
