+++ src/components/SimpleLocationPicker.tsx
