--- src/pages/swal.js
@@ -1,4 +1,0 @@
-import Swal from 'sweetalert2';
-import 'sweetalert2/dist/sweetalert2.min.css';
-
-export default Swal;
